Added latency and overall metrics

+++ src/main/kotlin/me/stroiker/flexibleloadgenerator/domain/LoadProfile.kt
@@ -1,6 +1,6 @@
 package me.stroiker.flexibleloadgenerator.domain
 
 data class LoadProfile(
-    val segments: Collection<Long>,
-    val timeScale: Long
+    val segments: Collection<Int>,
+    val timeScale: Int
 )
