+++ src/main/kotlin/me/stroiker/flexibleloadgenerator/mvc/model/LoadGeneratorStartRequest.kt
@@ -1,6 +1,6 @@
 package me.stroiker.flexibleloadgenerator.mvc.model
 
 internal class LoadGeneratorStartRequest(
-    val segments: Map<Long, Long> = mapOf(),
-    val timeScale: Long? = null,
+    val segments: Map<Int, Int> = mapOf(),
+    val timeScale: Int? = null,
 )
